[opt] Make Stardist code DRY
- create abstract stardist class for the two concrete stardist variant
  classes, because most of the functionality is identical
