+++ src/utils/models/availableSegmentationModels.ts
@@ -3,8 +3,7 @@ import { Cellpose } from "./Cellpose";
 import { CocoSSD } from "./CocoSSD";
 import { FullyConvolutionalSegmenter } from "./FullyConvolutionalSegmenter";
 import { Glas } from "./Glas/Glas";
-import { StardistVHE } from "./StardistVHE";
-import { StardistFluo } from "./StardistFluo";
+import { StardistVHE, StardistFluo } from "./Stardist";
 
 export const availableSegmenterModels: Array<Segmenter> = [
   new FullyConvolutionalSegmenter(),
